Add team builder application with responsive design
This commit includes adding an entire new team builder application. It contains HTML, CSS, and JS files to build a UI that allows the user to add, delete, and display team members' details. The design of the web pages is responsive to various screen sizes. A .gitignore file has also been added to exclude unnecessary files.