Refactor CSS and remove debug code in JS
In this update, we've made CSS rules more efficient by combining similar rules for '#contacts a' and '#course_link'. This will reduce redundancy in our CSS and make it more streamlined for future changes. In the JS script, a console logger used for debugging has been removed, as it's unnecessary in the final version.