Add MongoDB integration with Mongoose
This commit introduces MongoDB usage via the Mongoose module. It restructures the server-side code to communicate with the MongoDB database instead of using an in-memory array for member management. A seeder script has been added to populate the database and the 'id' property in the member object has been replaced with MongoDB's '_id'. The client-side react components have also been updated to use '_id'. In addition, a Team schema has been introduced for future team functionality.